Fix GCC 13 errors

+++ CesiumGltf/include/CesiumGltf/PropertyArrayView.h
@@ -37,10 +37,10 @@ public:
       : _values{CesiumUtility::reintepretCastSpan<const ElementType>(buffer)} {}
 
   const ElementType& operator[](int64_t index) const noexcept {
-    return this->_values[index];
+    return this->_values[size_t(index)];
   }
 
-  int64_t size() const noexcept { return this->_values.size(); }
+  int64_t size() const noexcept { return int64_t(this->_values.size()); }
 
   auto begin() { return this->_values.begin(); }
   auto end() { return this->_values.end(); }
@@ -151,7 +151,8 @@ public:
     index += _bitOffset;
     const int64_t byteIndex = index / 8;
     const int64_t bitIndex = index % 8;
-    const int bitValue = static_cast<int>(_values[byteIndex] >> bitIndex) & 1;
+    const int bitValue =
+        static_cast<int>(_values[size_t(byteIndex)] >> bitIndex) & 1;
     return bitValue == 1;
   }
 
@@ -193,10 +194,12 @@ public:
         _size{size} {}
 
   std::string_view operator[](int64_t index) const noexcept {
-    const size_t currentOffset =
-        getOffsetFromOffsetsBuffer(index, _stringOffsets, _stringOffsetType);
+    const size_t currentOffset = getOffsetFromOffsetsBuffer(
+        size_t(index),
+        _stringOffsets,
+        _stringOffsetType);
     const size_t nextOffset = getOffsetFromOffsetsBuffer(
-        index + 1,
+        size_t(index + 1),
         _stringOffsets,
         _stringOffsetType);
     return std::string_view(
