+++ CesiumGltf/include/CesiumGltf/PropertyTransformations.h
@@ -85,14 +85,14 @@ PropertyArrayCopy<T> transformArray(
     const std::optional<PropertyArrayView<T>>& scale) {
   std::vector<T> result(static_cast<size_t>(value.size()));
   for (int64_t i = 0; i < value.size(); i++) {
-    result[i] = value[i];
+    result[size_t(i)] = value[i];
 
     if (scale) {
-      result[i] = applyScale<T>(result[i], (*scale)[i]);
+      result[size_t(i)] = applyScale<T>(result[size_t(i)], (*scale)[i]);
     }
 
     if (offset) {
-      result[i] = result[i] + (*offset)[i];
+      result[size_t(i)] = result[size_t(i)] + (*offset)[i];
     }
   }
 
@@ -108,14 +108,14 @@ PropertyArrayCopy<NormalizedType> transformNormalizedArray(
     const std::optional<PropertyArrayView<NormalizedType>>& scale) {
   std::vector<NormalizedType> result(static_cast<size_t>(value.size()));
   for (int64_t i = 0; i < value.size(); i++) {
-    result[i] = normalize<T>(value[i]);
+    result[size_t(i)] = normalize<T>(value[i]);
 
     if (scale) {
-      result[i] = result[i] * (*scale)[i];
+      result[size_t(i)] = result[size_t(i)] * (*scale)[i];
     }
 
     if (offset) {
-      result[i] = result[i] + (*offset)[i];
+      result[size_t(i)] = result[size_t(i)] + (*offset)[i];
     }
   }
 
@@ -129,14 +129,14 @@ PropertyArrayCopy<glm::vec<N, double>> transformNormalizedVecNArray(
     const std::optional<PropertyArrayView<glm::vec<N, double>>>& scale) {
   std::vector<glm::vec<N, double>> result(static_cast<size_t>(value.size()));
   for (int64_t i = 0; i < value.size(); i++) {
-    result[i] = normalize<N, T>(value[i]);
+    result[size_t(i)] = normalize<N, T>(value[i]);
 
     if (scale) {
-      result[i] = result[i] * (*scale)[i];
+      result[size_t(i)] = result[size_t(i)] * (*scale)[i];
     }
 
     if (offset) {
-      result[i] = result[i] + (*offset)[i];
+      result[size_t(i)] = result[size_t(i)] + (*offset)[i];
     }
   }
 
@@ -150,14 +150,15 @@ PropertyArrayCopy<glm::mat<N, N, double>> transformNormalizedMatNArray(
     const std::optional<PropertyArrayView<glm::mat<N, N, double>>>& scale) {
   std::vector<glm::mat<N, N, double>> result(static_cast<size_t>(value.size()));
   for (int64_t i = 0; i < value.size(); i++) {
-    result[i] = normalize<N, T>(value[i]);
+    result[size_t(i)] = normalize<N, T>(value[i]);
 
     if (scale) {
-      result[i] = applyScale<glm::mat<N, N, double>>(result[i], (*scale)[i]);
+      result[size_t(i)] =
+          applyScale<glm::mat<N, N, double>>(result[size_t(i)], (*scale)[i]);
     }
 
     if (offset) {
-      result[i] = result[i] + (*offset)[i];
+      result[size_t(i)] = result[size_t(i)] + (*offset)[i];
     }
   }
